fdt_edit

Struct NodeRefInterruptController

pub struct NodeRefInterruptController<'a> {
    pub node: NodeRefGen<'a>,
}

Interrupt controller node reference.

Provides specialized access to interrupt controller nodes and their properties.

Fields
node: NodeRefGen<'a>

The underlying generic node reference

Implementations

impl<'a> NodeRefInterruptController<'a>

pub fn try_from(node: NodeRefGen<'a>) -> Result<Self, NodeRefGen<'a>>

Attempts to create an interrupt controller reference from a generic node.

Returns Err with the original node if itโs not an interrupt controller.

pub fn interrupt_cells(&self) -> Option<u32>

Get #interrupt-cells value

This determines how many cells are needed to describe interrupts referencing this controller

pub fn interrupt_address_cells(&self) -> Option<u32>

Get #address-cells value (used for interrupt-map)

pub fn is_interrupt_controller(&self) -> bool

Check if this is an interrupt controller

pub fn compatibles(&self) -> Vec<&str>

Get compatible list